# Add sort and dedup logic in C++ to `getFramesAtIndices`